Renamed some models to correct name and changed some services

+++ WTSuccess.Domain/Models/Question.cs
@@ -9,7 +9,7 @@ namespace WTSuccess.Domain.Models
     public class Question : EntityBase
     {
         Dictionary<string, string> QuestionsAnswer { get; set; }
-        public int ExamId { get; set; }
+        public ulong ExamId { get; set; }
         public Exam Exam { get; set; }
     }
 }
